+++ document/config.go
@@ -3,30 +3,32 @@ package document
 import (
 	"fmt"
 	"reflect"
+	"strings"
 
 	"github.com/spf13/viper"
 )
 
 type Config struct {
-	InputFiles      []string `mapstructure:"input" yaml:"input"`
-	Sources         []string `mapstructure:"source" yaml:"source"`
-	OutputDir       string   `mapstructure:"output" yaml:"output"`
-	TestOutput      string   `mapstructure:"tests" yaml:"tests"`
-	RenderFormat    string   `mapstructure:"format" yaml:"format"`
-	UseExports      bool     `mapstructure:"exports" yaml:"exports"`
-	ShortLinks      bool     `mapstructure:"short-links" yaml:"short-links"`
-	ReportMissing   bool     `mapstructure:"report-missing" yaml:"report-missing"`
-	Strict          bool     `mapstructure:"strict" yaml:"strict"`
-	DryRun          bool     `mapstructure:"dry-run" yaml:"dry-run"`
-	CaseInsensitive bool     `mapstructure:"case-insensitive" yaml:"case-insensitive"`
-	Bare            bool     `mapstructure:"bare" yaml:"bare"`
-	TemplateDirs    []string `mapstructure:"templates" yaml:"templates"`
-	PreRun          []string `mapstructure:"pre-run" yaml:"pre-run"`
-	PreBuild        []string `mapstructure:"pre-build" yaml:"pre-build"`
-	PreTest         []string `mapstructure:"pre-test" yaml:"pre-test"`
-	PostTest        []string `mapstructure:"post-test" yaml:"post-test"`
-	PostBuild       []string `mapstructure:"post-build" yaml:"post-build"`
-	PostRun         []string `mapstructure:"post-run" yaml:"post-run"`
+	InputFiles      []string          `mapstructure:"input" yaml:"input"`
+	Sources         []string          `mapstructure:"source" yaml:"source"`
+	SourceURLs      map[string]string `mapstructure:"source-url" yaml:"source-url"`
+	OutputDir       string            `mapstructure:"output" yaml:"output"`
+	TestOutput      string            `mapstructure:"tests" yaml:"tests"`
+	RenderFormat    string            `mapstructure:"format" yaml:"format"`
+	UseExports      bool              `mapstructure:"exports" yaml:"exports"`
+	ShortLinks      bool              `mapstructure:"short-links" yaml:"short-links"`
+	ReportMissing   bool              `mapstructure:"report-missing" yaml:"report-missing"`
+	Strict          bool              `mapstructure:"strict" yaml:"strict"`
+	DryRun          bool              `mapstructure:"dry-run" yaml:"dry-run"`
+	CaseInsensitive bool              `mapstructure:"case-insensitive" yaml:"case-insensitive"`
+	Bare            bool              `mapstructure:"bare" yaml:"bare"`
+	TemplateDirs    []string          `mapstructure:"templates" yaml:"templates"`
+	PreRun          []string          `mapstructure:"pre-run" yaml:"pre-run"`
+	PreBuild        []string          `mapstructure:"pre-build" yaml:"pre-build"`
+	PreTest         []string          `mapstructure:"pre-test" yaml:"pre-test"`
+	PostTest        []string          `mapstructure:"post-test" yaml:"post-test"`
+	PostBuild       []string          `mapstructure:"post-build" yaml:"post-build"`
+	PostRun         []string          `mapstructure:"post-run" yaml:"post-run"`
 }
 
 func ConfigFromViper(v *viper.Viper) (*Config, error) {
@@ -53,7 +55,8 @@ func (c *Config) check(v *viper.Viper) error {
 		}
 	}
 	for _, key := range v.AllKeys() {
-		if _, ok := fields[key]; !ok {
+		parts := strings.Split(key, ".")
+		if _, ok := fields[parts[0]]; !ok {
 			return fmt.Errorf("unknown field '%s' in config file", key)
 		}
 	}