+++ format/hugo_test.go
@@ -36,7 +36,7 @@ func TestHugoToLinkPath(t *testing.T) {
 
 func TestHugoProcessMarkdown(t *testing.T) {
 	form := Hugo{}
-	templ, err := document.LoadTemplates(&form)
+	templ, err := document.LoadTemplates(&form, "")
 	assert.Nil(t, err)
 
 	proc := document.NewProcessor(nil, &form, templ, &document.Config{})
@@ -95,24 +95,4 @@ func TestHugoCreateDirs(t *testing.T) {
 		"docs/src/_index.md",
 		"docs/test",
 	})
-}
-
-func TestGetGitOrigin(t *testing.T) {
-	conf, err := getGitOrigin("docs")
-
-	assert.Nil(t, err)
-	assert.Equal(t, conf.Repo, "https://github.com/mlange-42/modo")
-	assert.Equal(t, conf.Title, "modo")
-	assert.Equal(t, conf.Pages, "https://mlange-42.github.io/modo/")
-	assert.Equal(t, conf.Module, "github.com/mlange-42/modo/docs")
-}
-
-func TestRepoToTitleAndPages(t *testing.T) {
-	title, pages := repoToTitleAndPages("https://github.com/user/repo")
-	assert.Equal(t, title, "repo")
-	assert.Equal(t, pages, "https://user.github.io/repo/")
-
-	title, pages = repoToTitleAndPages("https://gitlab.com/user/repo")
-	assert.Equal(t, title, "repo")
-	assert.Equal(t, pages, "https://repo.com")
 }